requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
mysql-connector-python==8.2.0
pandas==2.1.3
python-dotenv==1.0.0
openpyxl==3.1.2
